requirements.txt
Click==7.0
click-log==0.3.2
jupyter==1.0.0
pandas==0.23.4
Pillow==5.3.0
requests==2.20.1
requests-cache==0.4.13
scispacy==0.2.2
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.0/en_core_sci_sm-0.2.0.tar.gz
joblib==0.13.2
scikit-learn==0.21.2
nltk==3.4.1